src/utils/automatic_setup/requirements.txt
psycopg2
requests
pytest
python-dotenv
requests-mock
jason
pgvector
uvicorn
fastapi